requirements.txt
torch==2.6.0
numpy==1.23.5
ffmpeg-python
loguru
pandas
boto3
ultralytics==8.3.112
pycocotools
tabulate
prettytable
lap==0.5.12
onnx==1.15.0
onnxruntime==1.16.1